test: move PC endpoint unit tests
